refactor(mc-scripts): move custom vite plugins to separate folder
