+++ packages/mc-scripts/src/commands/build-vite.ts
@@ -3,10 +3,10 @@ import path from 'path';
 import { build, type Plugin } from 'vite';
 import pluginGraphql from '@rollup/plugin-graphql';
 import pluginReact from '@vitejs/plugin-react';
-import pluginSvgr from 'vite-plugin-svgr';
 import { generateTemplate } from '@commercetools-frontend/mc-html-template';
 import { packageLocation as applicationStaticAssetsPath } from '@commercetools-frontend/assets';
 import paths from '../config/paths';
+import pluginSvgr from '../vite-plugins/vite-plugin-svgr';
 
 async function run() {
   const DEFAULT_PORT = parseInt(String(process.env.HTTP_PORT), 10) || 3001;
@@ -53,8 +53,6 @@ async function run() {
         },
       }),
       pluginSvgr({
-        include: '**/*.react.svg',
-        exportAsDefault: true,
         svgrOptions: {
           icon: false,
           svgoConfig: {
